moved mock method classes to conftest
